% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvi.R
\name{cvi}
\alias{cvi}
\title{function to compute CVI indexes (both I-CVI and S-CVI) for content
validity evaluation of a questionnaire/scale}
\usage{
cvi(x, strict = TRUE)
}
\arguments{
\item{x}{a data.frame of dummy (or factor) with 0 (first level) as 
"not relevant" and 1 (second level) "as relevant"}

\item{strict}{if TRUE, for I-CVI NA are treated like "not relevant"}
}
\description{
function to compute CVI indexes (both I-CVI and S-CVI) for content
validity evaluation of a questionnaire/scale
}
\references{
Polit, Beck, "The Content Validity Index: are you
sure you know whats' being reported? Critique and recommendations"
}
