% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbquest-package.R
\docType{package}
\name{lbquest-package}
\alias{lbquest}
\alias{lbquest-package}
\title{lbquest: lbquest}
\description{
Tools to help with cross cultural adaptation
    and pychometric validation of questionnaires/research instruments
}
\author{
\strong{Maintainer}: Luca Braglia \email{lbraglia@gmail.com}

}
\keyword{internal}
